Move some testing files from /mocha to /node
Summary:
since they are no longer used exclusively from mocha

Test Plan:
make test && make jest

Auditors: joel
