+++ Makefile
@@ -54,7 +54,7 @@ lint:
 	~/Khan/devtools/khan-linter/runlint.py
 
 test:
-	find -E src -type f -regex '.*/__tests__/.*\.jsx?' | xargs ./node_modules/.bin/mocha --reporter spec -r mocha/environment.js
+	find -E src -type f -regex '.*/__tests__/.*\.jsx?' | xargs ./node_modules/.bin/mocha --reporter spec -r node/environment.js
 
 build/ke.js:
 	(cd ke && ../node_modules/.bin/r.js -o requirejs.config.js out=../build/ke.js)
