+++ node/ke-deps-shim.js
